userpage page 1 changed adminpage all function working only user type combobox code remaining. alert message code added and working.

+++ source/UserPage.py
@@ -23,6 +23,7 @@ from pynmea import nmea
 import json
 from source.ThreadSound import ThreadingExample
 
+
 class UserMain(QtWidgets.QMainWindow):
     def __init__(self):
         QtWidgets.QMainWindow.__init__(self)
@@ -31,20 +32,38 @@ class UserMain(QtWidgets.QMainWindow):
         self.out=None
         self.Vheight=640
         self.Vwidth=480
-        self.flag=-1       #Flag for 1 -recording  -1 for no recodring and 0 for stop recording
-        #s
+        self.flag=-1 #Flag for 1 -recording  -1 for no recodring and 0 for stop recording
+        self.sliderFlag=-1
         #self.engine = pyttsx3.init()
         self.DisplayDateList()
         self.resultList=[]
+
         self.frameCount=1
+        #self.capture=None
         # elf.ui.webView.load(QtCore.QUrl("D:/Project/HawkEye/html/map.html"))
         self.fourcc=None
+        self.pauseFlag=0
+        self.defaultOpFolder="D:\\Project\\HawkEye\\video"
         self.outDir="D:\\Project\\HawkEye\\video"
-        self.ui.label_opdirtext.setText(self.outDir)
+        self.currentLatitude=""
+        self.currentLongitude=""
+        self.locationFlag=False
+        self.defaultGpsIp='127.0.0.1'
+        self.defaultGpsPort='14551'
+        self.VideoFlag=False
+        self.StreamFlag=False
+        self.ui.lineEdit_gpsIp.setText(self.defaultGpsIp)
+        self.ui.lineEdit_3.setText(self.defaultGpsPort)
+        self.ui.label_AlgoInit.setText("Check: NO")
+        self.ui.label_gpsStatus.setText("Not Connected")
+        self.gpsFlag=False
+        self.algoFlag=False
+        self.ui.label_outputDir.setText(self.defaultOpFolder)
+        self.ui.label_opdirtext.setText(self.defaultOpFolder)
         #self.ui.label_player.setPixmap(QPixmap("D:\\Project\\HawkEye\\resource\\images\\playerbackground.png"))
         self.setEnabledButtonOnCreate()
         #self.updateTable()
-        self.algo_init()
+        self.timer = QTimer()
         self.ButtonConnectOnCreate()
         self.ButtonDConnectOnCreate()
 
@@ -54,22 +73,29 @@ class UserMain(QtWidgets.QMainWindow):
 
 
 
-    def run(self):
-        while True:
-            # Do something
-            print('Doing something imporant in the background')
-            engine = pyttsx3.init()
-            engine.say("Human Detected")
-            engine.runAndWait()
-            time.sleep(1)  #Sleep Interval
+
+
+
+
 
 
     def algo_init(self):
-        options = {"model": "D:/Project/HawkEye/resource/darkflow/cfg/tiny-yolo-voc.cfg",
-                   "load": "D:/Project/HawkEye/resource/darkflow/bin/tiny-yolo-voc.weights",
-                   "threshold": 0.5, "gpu": 0.5}
-        self.tfnet = TFNet(options)
-        self.colors = [tuple(255 * np.random.rand(3)) for _ in range(10)]
+
+        if not(self.algoFlag):
+            options = {"model": "D:/Project/HawkEye/resource/darkflow/cfg/tiny-yolo-voc.cfg",
+                       "load": "D:/Project/HawkEye/resource/darkflow/bin/tiny-yolo-voc.weights",
+                       "threshold": 0.5, "gpu": 0.5}
+            self.tfnet = TFNet(options)
+            self.colors = [tuple(255 * np.random.rand(3)) for _ in range(10)]
+            self.ui.label_AlgoInit.setText("Check: Yes")
+            msg = QtWidgets.QMessageBox()
+            msg.about(self, "Message", "Algorithm Initialized")
+            self.ui.label_AlgoInit.setStyleSheet("background-color: rgb(170, 255, 127);\n"
+                                              "font: 10pt \"MS Shell Dlg 2\";")
+            self.algoFlag=True
+            return True
+        else:
+            return True
 
     def ButtonDConnectOnCreate(self):
         self.ui.tableWidget_Recording.doubleClicked.connect(self.on_DClickTable)
@@ -85,6 +111,46 @@ class UserMain(QtWidgets.QMainWindow):
         self.ui.pushButton_vstop.clicked.connect(self.vstop)
         self.ui.pushButton_show.clicked.connect(self.DisplayTable)
         self.ui.pushButton_today.clicked.connect(self.Today)
+        self.ui.horizontalSlider.valueChanged.connect(self.valuechange)
+        self.ui.horizontalSlider.sliderPressed.connect(self.SliderPressed)
+        self.ui.horizontalSlider.sliderReleased.connect(self.SliderReleased)
+        self.ui.horizontalSlider.sliderMoved.connect(self.SliderMoved)
+        self.ui.pushButton_pause.clicked.connect(self.pause)
+        self.ui.pushButton_ChangeOpDir.clicked.connect(self.ChangeOutputFolder)
+        self.ui.pushButton_FileDialog.clicked.connect(self.ChangeOutputFolder)
+        self.ui.pushButton_AlgoInit.clicked.connect(self.algo_init)
+        self.ui.pushButton_GpsCon.clicked.connect(self.connectGps)
+        self.ui.pushButton_logout.clicked.connect(self.OpenLoginPage)
+        self.ui.pushButton_SendAlert.clicked.connect(self.SendAlert)
+
+    def SendAlert(self):
+        if(self.locationFlag):
+            from firebase import firebase
+            firebase = firebase.FirebaseApplication('https://my-travel-junction.firebaseio.com/Location', None)
+            firebase.put('/LocationUpdate', "latitude", self.currentLatitude)
+            firebase.put('/LocationUpdate', "longitude", self.currentLongitude)
+            self.ui.label_5.setText("Location Sent:" + self.currentLatitude+ "," + self.currentLongitude)
+            msg = QtWidgets.QMessageBox()
+            msg.about(self, "Message", "Current Location Sent")
+
+        else:
+            msg = QtWidgets.QMessageBox()
+            msg.about(self, "Message", "Currrent Location Not Available")
+
+
+
+
+    def ChangeOutputFolder(self):
+        self.outDir=self.getOutputDir()
+        if self.outDir :
+            self.ui.label_outputDir.setText(self.outDir)
+            self.ui.label_opdirtext.setText(self.outDir)
+            msg = QtWidgets.QMessageBox()
+            msg.about(self, "Message", "Output folder changed")
+        else:
+            self.ui.label_outputDir.setText(self.defaultOpFolder)
+            self.ui.label_opdirtext.setText(self.defaultOpFolder)
+            self.outDir=self.defaultOpFolder
 
 
     def setEnabledButtonOnCreate(self):
@@ -94,7 +160,7 @@ class UserMain(QtWidgets.QMainWindow):
         self.ui.pushButton_play.setEnabled(False)
         self.ui.pushButton_start.setEnabled(True)
         self.ui.pushButton_vstop.setEnabled(False)
-        #self.ui.pushButton_prev.setEnabled(False)
+        self.ui.pushButton_pause.setEnabled(False)
        #self.ui.pushButton_next.setEnabled(False)
 
 
@@ -107,16 +173,19 @@ class UserMain(QtWidgets.QMainWindow):
 #webcam  method
     def start(self):
         self.flag=-1
-
+        self.StreamFlag=True
         if not(self.connectGps()):
+            self.ui.label_msg.setText("Streaming Stopped")
+            return
+        if not(self.algo_init()):
             self.ui.label_msg.setText("Streaming Stopped")
             return
 
         self.ui.pushButton_startRec.setEnabled(True)
         self.ui.pushButton_stop.setEnabled(True)
 
-        self.capture = cv2.VideoCapture(0) #0 for internal webcam 1 for external camera
-        if not  self.capture.isOpened():
+        self.captureCamera = cv2.VideoCapture(0) #0 for internal webcam 1 for external camera
+        if not  self.captureCamera.isOpened():
             msg = QtWidgets.QMessageBox()
             msg.about(self, "Error", "Video Device Not Connected")
             self.capture.release()
@@ -128,17 +197,19 @@ class UserMain(QtWidgets.QMainWindow):
             self.ui.label_msg.setText("Streaming Stopped")
             return
         else:
-            self.capture.set(cv2.CAP_PROP_FRAME_HEIGHT, 480)
-            self.capture.set(cv2.CAP_PROP_FRAME_WIDTH, 640)
+            self.captureCamera.set(cv2.CAP_PROP_FRAME_HEIGHT, 480)
+            self.captureCamera.set(cv2.CAP_PROP_FRAME_WIDTH, 640)
             self.timer =QTimer()
             self.timer.timeout.connect(self.refresh_frame)
-            self.timer.start(5)
+            self.timer.start(1)
             self.ui.label_msg.setText("Streaming...")
 
     def stop(self):
         self.timer.stop()
-        self.capture.release()
+        self.captureCamera.release()
         cv2.destroyAllWindows()
+        self.gpsFlag=False
+        self.StreamFlag=False
         self.disonnectGps()
         if(self.flag==1):
             self.jsonDataWrite()
@@ -206,7 +277,7 @@ class UserMain(QtWidgets.QMainWindow):
         dataResult = {}  # To store label and confidence
         finalData = {}  #To Store gps and mdataresult
         newFinalData = {}# To store final data and framenos
-        ret, self.image = self.capture.read()
+        ret, self.image = self.captureCamera.read()
         self.image = cv2.flip(self.image, 1)
         #IMAGE PROCESSING CODE GOES HERE
         frame = self.image
@@ -242,9 +313,8 @@ class UserMain(QtWidgets.QMainWindow):
                 gpsdata = self.mySocket.recv(1024).decode()
                 self.gpgga.parse(gpsdata)
                 latitude = self.gpgga.latitude
-                # lat_dir = self.gpgga.lat_direction
                 longitude = self.gpgga.longitude
-                # long_dir = self.gpgga.lon_direction
+
             #Write gps data to resultList
             if(self.flag==1):
                 finalData["latitude"] = latitude
@@ -253,6 +323,10 @@ class UserMain(QtWidgets.QMainWindow):
                 newFinalData[self.frameCount] = finalData
                 if(mdataResult):
                     self.resultList.append(newFinalData) #Appending Result to List
+
+            self.currentLatitude=latitude
+            self.currentLongitude=longitude
+            self.locationFlag=True
             self.displayPrediction(self.frameCount, latitude, longitude, outputtext)  # To Display Result
             self.frameCount += 1
         else :
@@ -294,18 +368,26 @@ class UserMain(QtWidgets.QMainWindow):
         self.ui.label_player.setScaledContents(True)
 
     def connectGps(self):
-        self.host = '127.0.0.1'
-        self.port = 14551
-        self.mySocket = socket.socket()
-        try:
-            self.mySocket.connect((self.host, self.port))
-            print("gps connected")
+        if not (self.gpsFlag):
+            self.host =self.ui.lineEdit_gpsIp.text() #127.0.0.1
+            self.port = self.ui.lineEdit_3.text() #14551
+            self.mySocket = socket.socket()
+            try:
+                self.mySocket.connect((self.host, int(self.port)))
+                self.ui.label_gpsStatus.setText("Connected")
+                msg = QtWidgets.QMessageBox()
+                msg.about(self, "Success", "Gps Connected")
+                self.ui.label_gpsStatus.setStyleSheet("background-color: rgb(170, 255, 127);\n"
+                                                   "font: 10pt \"MS Shell Dlg 2\";")
+                self.gpsFlag=True
+                return True
+            except:
+                msg = QtWidgets.QMessageBox()
+                msg.about(self, "Error", "Gps Not Connected")
+                self.gpsFlag=False
+                return False
+        else:
             return True
-        except:
-            msg = QtWidgets.QMessageBox()
-            msg.about(self, "Error", "Gps Not Connected")
-            return False
-
 
     def disonnectGps(self):
         self.mySocket.close()
@@ -324,28 +406,62 @@ class UserMain(QtWidgets.QMainWindow):
         self.db_Cursor = self.db_Object.cursor()
 
     def play(self, filename):
+
+        # if(self.pauseFlag==0):
+        self.pauseFlag=1
         self.flag = -1
+        self.VideoFlag=True
         self.capture = cv2.VideoCapture(filename)
+        frameCount=self.capture.get(cv2.CAP_PROP_FRAME_COUNT)
+        self.sliderValue=1
+        self.sliderFlag=0
+        self.Trackbar(frameCount)
+
         self.capture.set(cv2.CAP_PROP_FRAME_HEIGHT, 480)
         self.capture.set(cv2.CAP_PROP_FRAME_WIDTH, 640)
-        self.timer = QTimer()
         self.timer.timeout.connect(self.refresh_frame2)
-        self.timer.start(100)
+        self.timer.start(100)  #time as argument
+    # if(self.pauseFlag==1):
+        #     self.timer.timeout.connect(self.refresh_frame2)
+        #     self.timer.start(100)
+        #     #To Start Playing a paused video
+
+    def pause(self):
+        self.pauseFlag = 1
+        self.timer.stop()
+        #self.ButtonConnectOnCreate()
+
+
+
+
+    def Timer(self,time):
+
+        self.timer.timeout.connect(self.refresh_frame2)
+        self.timer.start(time)
 
     def vstop(self):
-        self.timer.stop()
-        self.capture.release()
-        cv2.destroyAllWindows()
-        self.setEnabledButtonOnCreate()
-        self.ui.label_player_2.setPixmap(QPixmap("D:\\Project\\HawkEye\\resource\\images\\playerbackground.png"))
+
+        if (self.capture.isOpened()):
+            self.timer.stop()
+            self.VideoFlag=False
+            self.capture.release()
+            cv2.destroyAllWindows()
+            self.setEnabledButtonOnCreate()
+            self.ui.label_player_2.setPixmap(QPixmap("D:\\Project\\HawkEye\\resource\\images\\playerbackground.png"))
 
     def refresh_frame2(self):
         ret, self.image = self.capture.read()
+        if(self.sliderFlag==0):
+            self.ui.horizontalSlider.setValue(self.sliderValue)
+            self.ui.label_frameno.setText(str(self.sliderValue))
+
         if ret:
         # IMAGE PROCESSING CODE GOES HERE
             self.displayImage2(self.image)
         else:
             self.vstop()
+
+        self.sliderValue+=1
 
     def displayImage2(self, img):
         qFormat = QImage.Format_Indexed8
@@ -361,7 +477,28 @@ class UserMain(QtWidgets.QMainWindow):
         self.ui.label_player_2.setPixmap(QPixmap.fromImage(outImage))
         self.ui.label_player_2.setScaledContents(True)
 
+    def Trackbar(self,maxFrame):
+        self.ui.horizontalSlider.setMinimum(1)
+        self.ui.horizontalSlider.setMaximum(maxFrame)
+        self.ui.horizontalSlider.setValue(1)
+        self.ui.horizontalSlider.setTickPosition(QtWidgets.QSlider.TicksBelow)
+        self.ui.horizontalSlider.setTickInterval(1)
 
+    def valuechange(self):
+        if(self.capture.isOpened()):
+            frameNo = self.ui.horizontalSlider.value()
+
+            self.capture.set(cv2.CAP_PROP_POS_FRAMES, frameNo)
+
+
+    def SliderReleased(self):
+        self.sliderFlag=0
+
+    def SliderPressed(self):
+        self.sliderFlag=1
+
+    def SliderMoved(self):
+        self.sliderFlag=1
 
 ##################
     def on_DClickTable(self):
@@ -378,7 +515,7 @@ class UserMain(QtWidgets.QMainWindow):
 
         for currentQTableWidgetItem in self.ui.tableWidget_Recording.selectedItems():
             self.currentFile =self.outDir+"/"+currentQTableWidgetItem.text()
-            print(self.currentFile)
+            #print(self.currentFile)
             self.play(self.currentFile)
 
     def DisplayTable(self):
@@ -399,15 +536,21 @@ class UserMain(QtWidgets.QMainWindow):
             fileName=row[0]
             date_info = row[1]
             filepath = self.outDir + "/" + fileName
-            filesize = os.path.getsize(filepath) / (1024 * 1024)
-            filesize = '{:.0f}'.format(filesize)
-            filesize = filesize + "Mb"
-
-            self.ui.tableWidget_Recording.insertRow(numRows)
-            self.ui.tableWidget_Recording.setItem(numRows,0, QtWidgets.QTableWidgetItem(fileName))
-            self.ui.tableWidget_Recording.setItem(numRows, 1, QtWidgets.QTableWidgetItem(date_info))
-            self.ui.tableWidget_Recording.setItem(numRows, 2, QtWidgets.QTableWidgetItem(filesize))
-            numRows=numRows+1
+            if(os.path.exists(filepath)):
+                filesize = os.path.getsize(filepath) / (1024 * 1024)
+                filesize = '{:.0f}'.format(filesize)
+                filesize = filesize + "Mb"
+                videoCap = cv2.VideoCapture(filepath)
+                self.ui.tableWidget_Recording.insertRow(numRows)
+                self.ui.tableWidget_Recording.setItem(numRows,0, QtWidgets.QTableWidgetItem(fileName))
+                self.ui.tableWidget_Recording.setItem(numRows, 1, QtWidgets.QTableWidgetItem(date_info))
+                self.ui.tableWidget_Recording.setItem(numRows, 2, QtWidgets.QTableWidgetItem(filesize))
+                self.ui.tableWidget_Recording.setItem(numRows, 3, QtWidgets.QTableWidgetItem(str(videoCap.get(cv2.CAP_PROP_FRAME_COUNT))))
+                numRows=numRows+1
+                videoCap.release()
+            else :
+                msg = QtWidgets.QMessageBox()
+                msg.about(self, "Error", "File not found")
 
         self.DBDisconnect()
 
@@ -444,11 +587,14 @@ class UserMain(QtWidgets.QMainWindow):
             filesize = os.path.getsize(filepath) / (1024 * 1024)
             filesize = '{:.0f}'.format(filesize)
             filesize = filesize + "Mb"
+            videoCap = cv2.VideoCapture(filepath)
             self.ui.tableWidget_Recording.insertRow(numRows)
             self.ui.tableWidget_Recording.setItem(numRows, 0, QtWidgets.QTableWidgetItem(fileName))
             self.ui.tableWidget_Recording.setItem(numRows, 1, QtWidgets.QTableWidgetItem(date_info))
             self.ui.tableWidget_Recording.setItem(numRows, 2, QtWidgets.QTableWidgetItem(filesize))
+            self.ui.tableWidget_Recording.setItem(numRows, 3, QtWidgets.QTableWidgetItem(str(videoCap.get(cv2.CAP_PROP_FRAME_COUNT))))
             numRows = numRows + 1
+            videoCap.release()
 
     def DisplayDateList(self):
         self.ui.comboBox.clear()
@@ -502,6 +648,15 @@ class UserMain(QtWidgets.QMainWindow):
                     str(varDate.hour)+"."+str(varDate.minute)+"."+str(varDate.second)
         return newFileName
 
+    def OpenLoginPage(self):
+        if(self.VideoFlag):
+            self.vstop()
+        if(self.StreamFlag):
+            self.stop()
+        from source.LoginPage import LoginMain
+        self.close()
+        self.objectUserMain = LoginMain()
+        self.objectUserMain.show()
 
 
 if __name__ == "__main__":
